Reorganize and add new promoters of the event
Rename component 'sponsors' to 'promoters' to better represent the
different categories of institutions that are helping to make the event.

Change how sponsors, supporters and organizers are displayed in the
page. Put sponsors and supporters side by side in screen sizes bigger
than medium-sized.

Add logos of new sponsors and supporters.

Signed-off-by: Renato Cordeiro Ferreira <renato.cferreira@hotmail.com>