Fix conditional logic and refactoring
Remove unnecessary variables in `2018.1.vue` and add computed methods to
the page component. Also clean up code to make it easier to understand.

Setup the `Countdown.vue` component to be more generic, so that it can
be used for any event.

Add proper css to subscribe and rules buttons. Also fix conditionals in
which they are shown.

Signed-off-by: Victor Hugo M. Pinto <victor2142@gmail.com>